Show/hide/autoplay update...
Had autoplay earlier tonight, but there were a lot of exceptional scenarios that would also allow the player to auto play. We need to figure out a way to only have it auto play on the first time attempting to playback content when the queue is empty and the player is not initialized.

However, the following HAVE been implemented:

- Show/hide player
- Alert user to add content to queue if empty when trying to begin playback or show an unintialized player
- Alert user to begin playback if the queue is not empty but the player is uninitialized
- Alert user if a song ends and there is no more music in the queue
- Alert user if they try to fastforward to the next song but there is no more music in the queue (this previously caused an error in the player if they tried to poll an empty queue)

Not pushing to master yet because this is some awk functionality that MAY (or may not) require some refactoring.